+++ cliente/src/components/UserMenu.vue
@@ -1,9 +1,11 @@
 <template>
 <div>
-  <div class="row backgroundColor">
-    <div class="whiteText col-12">
+  <div class="row justify-content-center align-items-center backgroundColor">
+    <div class="col-10 text-center">
+      <h3 class="boldText whiteText">BIENVENIDO {{ name }}</h3>
+    </div>
+    <div class="col-2 elementPosition">
       <a href="#" id="logOut" class="boldText whiteText" v-on:click="logOut">LOGOUT</a>
-      <p class="boldText text-center">BIENVENIDO {{ name }}</p>
     </div>
   </div>
   <b-navbar toggleable="md" class="row backgroundColor">
@@ -38,22 +40,25 @@ export default {
     // formEditRequisition
   },
   mounted() {
-    axios.get('http://127.0.0.1:8000/users/' + sessionStorage.getItem('idUser'), {
-        //  params: {
-        //   id: sessionStorage.getItem('idUser')
-        // },
-        headers: {
-          'Authorization': 'JWT' + sessionStorage.getItem('idToken')
-        }
-      })
-      .then((response) => {
-        this.name = response.data.username;
-      })
-      .catch((error) => {
-        console.log("No salio");
-      });
+    this.load_username()
   },
   methods: {
+    load_username() {
+      axios.get('http://127.0.0.1:8000/users/' + sessionStorage.getItem('idUser'), {
+          //  params: {
+          //   id: sessionStorage.getItem('idUser')
+          // },
+          headers: {
+            'Authorization': 'JWT' + sessionStorage.getItem('idToken')
+          }
+        })
+        .then((response) => {
+          this.name = response.data.username;
+        })
+        .catch((error) => {
+          console.log("No salio");
+        });
+    },
     new_requirement: function() {
       this.requirementSection = !this.requirementSection;
       this.errorSection = false;
@@ -65,9 +70,23 @@ export default {
       this.requirementSection = false;
     },
     logOut: function() {
-      // axios.get('http://127.0.0.1:8000/rest-auth/logout/');
-      sessionStorage.clear();
-      this.$router.push('/');
+      this.$swal({
+        title: '¿Seguro que quieres salir?',
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#2699FB',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, estoy seguro',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+        if (result.value) {
+          this.$swal(
+            'Cierre de sesion exitoso',
+          )
+          sessionStorage.clear();
+          this.$router.push('/');
+        }
+      })
     }
   },
   created() {
